+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/CryptoPublicKey.kt
@@ -22,17 +22,10 @@ import at.asitplus.crypto.datatypes.asn1.readOid
 import at.asitplus.crypto.datatypes.asn1.runRethrowing
 import at.asitplus.crypto.datatypes.io.ByteArrayBase64Serializer
 import at.asitplus.crypto.datatypes.io.MultiBase
-import at.asitplus.crypto.datatypes.misc.ANSI_COMPRESSED_PREFIX_1
-import at.asitplus.crypto.datatypes.misc.ANSI_COMPRESSED_PREFIX_2
-import at.asitplus.crypto.datatypes.misc.ANSI_UNCOMPRESSED_PREFIX
-import at.asitplus.crypto.datatypes.misc.UVarInt
-import at.asitplus.crypto.datatypes.misc.compressY
-import at.asitplus.crypto.datatypes.misc.decompressY
-import at.asitplus.crypto.datatypes.misc.toUInt
+import at.asitplus.crypto.datatypes.misc.*
+import com.ionspin.kotlin.bignum.integer.Sign
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-
-typealias Signum = Boolean
 
 /**
  * Representation of a public key structure
@@ -91,16 +84,18 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
             val bytes = multiKeyRemovePrefix(input)
             val decoded = MultiBase.decode(bytes)
             val codec = UVarInt.fromByteArray(decoded.sliceArray(0..1)).toULong()
+            val keyBytes = decoded.copyOfRange(2, decoded.size)
 
             return when (codec) {
-                0x1200uL, 0x1201uL, 0x1202uL ->
-                    EC.fromAnsiX963Bytes(decoded.drop(2).toByteArray())
-
-                0x1290uL, 0x1291uL, 0x1292uL ->
-                    EC.fromAnsiX963Bytes(decoded.drop(2).toByteArray())
+                0x1200uL, 0x1290uL ->
+                    EC.fromAnsiX963Bytes(ECCurve.SECP_256_R_1, keyBytes)
+                0x1201uL, 0x1291uL ->
+                    EC.fromAnsiX963Bytes(ECCurve.SECP_384_R_1, keyBytes)
+                0x1202uL, 0x1292uL ->
+                    EC.fromAnsiX963Bytes(ECCurve.SECP_521_R_1, keyBytes)
 
                 0x1205uL ->
-                    Rsa.fromPKCS1encoded(decoded.drop(2).toByteArray())
+                    Rsa.fromPKCS1encoded(keyBytes)
 
                 else ->
                     throw IllegalArgumentException("Unknown public key identifier $codec")
@@ -121,7 +116,7 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
                         ?: throw Asn1Exception("Curve not supported: $curveOid")
 
                     val bitString = (src.nextChild() as Asn1Primitive).readBitString()
-                    if (bitString.rawBytes.first() != ANSI_UNCOMPRESSED_PREFIX) throw Asn1Exception("EC key not prefixed with 0x04")
+                    if (bitString.rawBytes.first() != ANSIECPrefix.UNCOMPRESSED.prefixByte) throw Asn1Exception("EC key not prefixed with 0x04")
                     val xAndY = bitString.rawBytes.drop(1)
                     val coordLen = curve.coordinateLengthBytes.toInt()
                     val x = xAndY.take(coordLen).toByteArray()
@@ -150,7 +145,16 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
         @Throws(Throwable::class)
         fun fromIosEncoded(it: ByteArray): CryptoPublicKey =
             when (it[0].toUByte()) {
-                ANSI_UNCOMPRESSED_PREFIX.toUByte() -> EC.fromAnsiX963Bytes(it)
+                ANSIECPrefix.UNCOMPRESSED.prefixUByte -> {
+                    val curve = when (it.size) {
+                        /** apple does not encode the curve identifier, but it is implied as one of the ios supported curves */
+                        65 -> ECCurve.SECP_256_R_1
+                        97 -> ECCurve.SECP_384_R_1
+                        133 -> ECCurve.SECP_521_R_1
+                        else -> throw IllegalArgumentException("Unknown curve in iOS raw key")
+                    }
+                    EC.fromAnsiX963Bytes(curve, it)
+                }
                 DERTags.DER_SEQUENCE -> Rsa.fromPKCS1encoded(it)
                 else -> throw IllegalArgumentException("Unsupported Key type")
             }
@@ -295,9 +299,7 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
     @Serializable
     @SerialName("EC")
     data class EC private constructor(
-        val curve: ECCurve,
-        @Serializable(with = ByteArrayBase64Serializer::class) val x: ByteArray,
-        @Serializable(with = ByteArrayBase64Serializer::class) val y: ByteArray,
+        val publicPoint: ECPoint,
         var useCompressedRepresentation: Boolean,
     ) : CryptoPublicKey() {
 
@@ -307,8 +309,15 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
         constructor(
             curve: ECCurve,
             x: ByteArray,
-            yIndicator: Signum,
-        ) : this(curve, x, decompressY(curve, x, yIndicator), true)
+            sign: Sign,
+        ) : this(ECPoint.fromCompressed(curve, x, sign), true)
+
+        /** Constructor for compressed keys */
+        constructor(
+            curve: ECCurve,
+            x: ByteArray,
+            usePositiveY: Boolean,
+        ) : this(ECPoint.fromCompressed(curve, x, usePositiveY), true)
 
         /**
          * Constructor for uncompressed keys
@@ -317,7 +326,14 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
             curve: ECCurve,
             x: ByteArray,
             y: ByteArray,
-        ) : this(curve, x, y, false)
+        ) : this(ECPoint.fromUncompressed(curve, x, y), false)
+
+        val curve get() = publicPoint.curve
+        val x get() = publicPoint.x
+        val xBytes get() = publicPoint.xBytes
+        val y get() = publicPoint.y
+        val yBytes get() = publicPoint.yBytes
+        val yCompressed get() = publicPoint.yCompressed
 
         override val oid = EC.oid
 
@@ -327,18 +343,16 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
         fun toAnsiX963Encoded(): ByteArray =
             when (useCompressedRepresentation) {
                 true -> {
-                    val prefix = (2U + compressY().toUInt()).toByte()
-                        .also { require(it == ANSI_COMPRESSED_PREFIX_1 || it == ANSI_COMPRESSED_PREFIX_2) }
                     byteArrayOf(
-                        prefix,
-                        *x.ensureSize(curve.coordinateLengthBytes)
+                        ANSIECPrefix.forSign(yCompressed).prefixByte,
+                        *xBytes
                     )
                 }
 
                 false -> byteArrayOf(
-                    ANSI_UNCOMPRESSED_PREFIX,
-                    *x.ensureSize(curve.coordinateLengthBytes),
-                    *y.ensureSize(curve.coordinateLengthBytes)
+                    ANSIECPrefix.UNCOMPRESSED.prefixByte,
+                    *xBytes,
+                    *yBytes
                 )
             }
 
@@ -358,10 +372,10 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
          * The keybytes are ANSI X9.63 encoded (important for compression)
          */
         override val didEncoded by lazy {
-            val codec = (0x12 shl 8).toUInt() + when (curve) {
-                ECCurve.SECP_256_R_1 -> 0x00u + 0x90u * (1U - useCompressedRepresentation.toUInt())
-                ECCurve.SECP_384_R_1 -> 0x01u + 0x90u * (1U - useCompressedRepresentation.toUInt())
-                ECCurve.SECP_521_R_1 -> 0x02u + 0x90u * (1U - useCompressedRepresentation.toUInt())
+            val codec = when (curve) {
+                ECCurve.SECP_256_R_1 -> if (useCompressedRepresentation) 0x1200u else 0x1290u
+                ECCurve.SECP_384_R_1 -> if (useCompressedRepresentation) 0x1201u else 0x1291u
+                ECCurve.SECP_521_R_1 -> if (useCompressedRepresentation) 0x1202u else 0x1292u
             }
             PREFIX_DID_KEY + ":" + MultiBase.encode(
                 MultiBase.Base.BASE58_BTC,
@@ -377,54 +391,34 @@ sealed class CryptoPublicKey : Asn1Encodable<Asn1Sequence>, Identifiable {
 
             other as EC
 
-            if (curve != other.curve) return false
-            if (!iosEncoded.contentEquals(other.iosEncoded)) return false
-
-            return true
+            return (this.publicPoint == other.publicPoint)
         }
 
-        override fun hashCode(): Int {
-            var result = curve.hashCode()
-            result = 31 * result + x.contentHashCode()
-            result = 31 * result + y.contentHashCode()
-            return result
-        }
+        override fun hashCode() =
+            publicPoint.hashCode()
 
         companion object : Identifiable {
-
-            private fun getCurve(coordSize: Int) = ECCurve.entries
-                .find { it.coordinateLengthBytes.toInt() == coordSize }
-                ?: throw IllegalArgumentException("Unknown Curve")
 
             /**
              * Decodes a key from its ANSI X9.63 representation
              */
             @Throws(Throwable::class)
-            fun fromAnsiX963Bytes(src: ByteArray): EC {
-                val curve: ECCurve
-                val numBytes: Int
-                val x: ByteArray
-                val y: ByteArray
+            fun fromAnsiX963Bytes(curve: ECCurve, src: ByteArray): EC {
+                val numBytes = curve.coordinateLength.bytes.toInt()
 
-                when (src[0]) {
-                    ANSI_UNCOMPRESSED_PREFIX -> {
-                        curve = getCurve((src.size - 1) / 2)
-                        numBytes = curve.coordinateLengthBytes.toInt()
-                        x = src.drop(1).take(numBytes).toByteArray()
-                        y = src.drop(1).drop(numBytes).take(numBytes).toByteArray()
-                    }
+                val prefix = runCatching { ANSIECPrefix.fromPrefixByte(src[0]) }
+                    .getOrElse { throw IllegalArgumentException("Invalid X9.63 EC key format") }
 
-                    ANSI_COMPRESSED_PREFIX_1, ANSI_COMPRESSED_PREFIX_2 -> {
-                        curve = getCurve(src.size - 1)
-                        numBytes = curve.coordinateLengthBytes.toInt()
-                        x = src.drop(1).take(numBytes).toByteArray()
-                        y = decompressY(curve, x, (src[0] - 2) == 1)
-                    }
-
-                    else -> throw IllegalArgumentException("Invalid X9.63 EC key format")
+                if (prefix.isUncompressed) {
+                    require(src.size == (2*numBytes + 1))
+                    val x = src.copyOfRange(1, numBytes+1)
+                    val y = src.copyOfRange(numBytes+1, 2*numBytes+1)
+                    return EC(curve, x, y)
+                } else {
+                    require(src.size == (numBytes + 1))
+                    val x = src.copyOfRange(1, src.size)
+                    return EC(curve, x, prefix.compressionSign)
                 }
-
-                return EC(curve = curve, x = x, y = y)
             }
 
             override val oid = KnownOIDs.ecPublicKey